Don't crash if the directory we are showing is deleted.
@@ -716,6 +716,12 @@ void KonqKfmIconView::slotNewItems( cons

 void KonqKfmIconView::slotDeleteItem( KFileItem * _fileitem )
 {
+    if ( _fileitem == m_dirLister->rootItem() )
+    {
+        m_pIconView->setRootItem( 0L );
+        return;
+    }
+

svn path=/trunk/kdebase/konqueror/; revision=141426

+++ konqueror/iconview/konq_iconview.cc
@@ -716,6 +716,12 @@ void KonqKfmIconView::slotNewItems( const KFileItemList& entries )
 
 void KonqKfmIconView::slotDeleteItem( KFileItem * _fileitem )
 {
+    if ( _fileitem == m_dirLister->rootItem() )
+    {
+        m_pIconView->setRootItem( 0L );
+        return;
+    }
+
     KonqDirPart::deleteItem( _fileitem );
 
     //kdDebug(1202) << "KonqKfmIconView::slotDeleteItem(...)" << endl;
